requirements.txt
django==4.2
djangorestframework==3.14.0
boto3==1.26.137
django-cors-headers==4.3.1
python-decouple==3.8
python-dotenv==1.0.0
pynamodb==5.3.4
drf-spectacular==0.27.0
moto==4.2.14
gunicorn==23.0.0

# Dependencias para notificaciones
twilio==8.2.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-mock==3.12.0
factory-boy==3.3.0